+++ CHANGELOG.md
@@ -80,6 +80,7 @@
 - Binance parser: added re-named operations for Statements. ([#292](https://github.com/BittyTax/BittyTax/issues/292))
 - OKX parser: added new export formats. ([#293](https://github.com/BittyTax/BittyTax/issues/293))
 - Exodus parser: added staking export.
+- Conversion tool: added warning if duplicate rows found.
 ### Changed
 - Binance parser: performance improvements for large data sets.
 - Coinbase Pro parser: performance improvements for large data sets.